requirements.txt
fastapi
uvicorn
tensorflow
python-multipart
pillow
huggingface_hub
pyhton-jose
pymongo
bcrypt
PyJWT>=2.0